refactor(viewer): decompose mode.rs God Module into 4 focused modules
mode.rs (3,652 LOC) → mode/ directory with 4 modules:
- mod.rs (1,650): state machine, debug controls, DOM utilities
- trpg_controls.rs (1,656): actor admin CRUD, new game flow, presets
- room_hub.rs (702): room management, localStorage persistence
- mcp_rpc.rs (303): MCP JSON-RPC layer, response parsing

Also fixes 10 compiler warnings (cfg-gated imports, unused variables).
Zero errors, zero warnings on native target.
